pyproject.toml
[project]
name = "trump-truth-social-feed"
version = "0.1.0"
description = "Daily LLM enrichment pipeline for Trump's Truth Social posts"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jinja2>=3.0",
    "litellm>=1.0",
    "pandas>=3.0.0",
    "python-dotenv>=1.0",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.14",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[project.scripts]
ttsfeed = "ttsfeed.pipeline:main"
